Change the function report order
It seems some unofficial title expect the function
reports are placed in a specific order, such as
sw input comes first, then coin input follows, etc.

Such titles compile a command series to query inputs
as in the same order as it they appear in the report,
then fail to parse the coin input coming later.

Version is changed to 1.10d.